DALA-5732: Change automatic reset behavior of "notify me immediately" flag
Implemented Review Comments (compact version of updateNotifyMeImmediatelyIfRequired, tests now realized in Utils Test Class)
