+++ gitinfo
@@ -1,0 +1,1 @@
+{"commit":"83e0939d33bead4c537c3c0482d8cae37703e217","time":"2025-04-27T09:58:46Z","branch":"main"}
